requirements.txt
# Core dependencies
streamlit>=1.28.0
numpy>=1.24.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0
pandas>=2.1.0
rich>=10.14.0

# API integrations
requests-cache>=1.1.0  # Optional for HTTP request caching
backoff>=2.2.1  # For API rate limiting and retries

# Concurrency
concurrent-log-handler>=0.9.24  # Thread-safe logging

# Optional - for development
black>=23.10.0
pylint>=3.0.0
pytest>=7.4.0
